split and reorganize the package in order to upload our code to the wechat public platform.
