+++ pages/mainPackage/createActivity/createActivity.wxml
@@ -12,7 +12,7 @@ bindregionchange="regionchange"
 show-location style="width: 100%; height:490rpx;"></map>
 
 <view class="solid-bottom colomn">
-  <image src="./images/wave.gif" 
+  <image src="../../../images/wave.gif" 
   mode="scaleToFill" class="gif-wave">
   </image>
 </view>
@@ -78,7 +78,7 @@ show-location style="width: 100%; height:490rpx;"></map>
   </view>
 </view>
 <view class="solid-bottom colomn">
-<image src="./images/wave.gif" mode="scaleToFill" class="gif-wave"></image>
+<image src="../../../images/wave.gif" mode="scaleToFill" class="gif-wave"></image>
 </view>
