+++ pages/mainPackage/createActivity/package.json
