+++ pages/mainPackage/exploreActivity/exploreActivity.json
