+++ pages/mainPackage/exploreActivity/exploreActivity.test.js
