+++ pages/mainPackage/exploreActivity/exploreActivity.wxml
@@ -3,7 +3,7 @@
     bindscrolltolower="bottomHandler" scrolltoupper="upperHandler" class="scrollPage">
 
 
-  <image src='./images/wave.gif' mode='widthFix' class='gif' style='width:100%;height:486rpx'></image>
+  <image src='../../../images/wave.gif' mode='widthFix' class='gif' style='width:100%;height:486rpx'></image>
   
   <view class='card-list'>
     <view bindtap="clickForActivityDetail" class="card-li bg-{{randomColorArr[key]}}" wx:for="{{activitylist}}" wx:key="key" wx:for-index="key" id="{{key}}">
