+++ pages/mainPackage/exploreActivity/exploreActivity.wxss
