+++ pages/mainPackage/exploreActivity/package.json
