--- pages/messageBoard/messageBoard.wxss
@@ -1,85 +1,0 @@
-
-.names {
-  display: flex;
-  font-size: 30rpx;
-  line-height: 40rpx;
-}
- 
-.input_null {
-  color: #c9c9c9;
-}
- 
-.replyAll {
-  position:absolute;
-}
- 
-.release {
-  align-items: flex-end; /*底部对齐*/
-  box-sizing: border-box;
-  position: fixed;
-  left: 0;
-  bottom: 0;
-  width: 100%;
-  padding: 18rpx 0 18rpx 30rpx;
-  background-color: #f7f8f7;
-  font-size: 28rpx;
-  z-index: 999;
-}
- 
-.replyinfo1{ 
-  display: flex;
-  justify-content: space-between; /*两端对齐*/
-  font-size: 35rpx;
-}
-.replyinfo2{ 
-  display: flex;
-  justify-content: space-between; /*两端对齐*/
-}
- 
-.release textarea {
-  width: 550rpx;
-  min-height: 34rpx;
-  max-height: 102rpx; /*最多显示三行*/
-  border-width: 15rpx 20rpx; /*使用padding与预期留白不一致，故使用border*/
-  border-style: solid;
-  border-color: #fff;
-  line-height: 34rpx;
-  font-size: 28rpx;
-  background-color: #fff;
-  border-radius: 4rpx;
-}
- 
-.release .text {
-  font-size: 40rpx;
-  color: #c9c9c9;
-}
- 
-.cancel {
-  width: 240rpx;
-  height: 64rpx;
-  line-height: 64rpx;
-  text-align: center;
-  color: #6c0;
-  margin: 0 3px;
-  padding: 0;
-}
- 
-.release .submit {
-  width: 120rpx;
-  height: 64rpx;
-  line-height: 64rpx;
-  text-align: center;
-  color: #6c0;
-  margin: 0 3px;
-  padding: 0;
-}
- 
-.pro-box .info .text .delete {
-  color: #f68135;
-  border-radius: 50rpx;
-  border: 1px solid #f68135;
-  font-size: 28 rpx;
-  width: 150rpx;
-  height: 48rpx;
-  text-align: center;
-}
