+++ pages/subPackage/activityInfo/activityInfo.js
