+++ pages/subPackage/activityInfo/activityInfo.json
