+++ pages/subPackage/activityInfo/activityInfo.test.js
