+++ pages/subPackage/activityInfo/activityInfo.wxml
