+++ pages/subPackage/activityInfo/activityInfo.wxss
