+++ pages/subPackage/activityInfo/package.json
