+++ pages/subPackage/initiatedActivity/initiatedActivity.js
@@ -1,61 +1,47 @@
-// pages/exploreActivity/exploreActivity.js
+// pages/initiatedActivity/initiatedActivity.js
 const app = getApp();
-var util = require('../../utils/util.js');
+var util = require('../../../utils/util.js');
 Page({
 
   /**
    * 页面的初始数据
    */
   data: {
-    fakeList: app.globalData.activity_list_fake,
+    newlist: app.globalData.activity_list_fake,
     
     randomColorArr: [], 
-    longitude: undefined,
-    latitude: undefined,
-    
     activitylist: [],
+    show_message: false,
     
     
     limit_per_request: 3,
     lastActivityTime: '', 
-    longitude: 116.0,       // 随便定的坐标，防止请求getActivityList()中请求位置出错
-    latitude: 40.0,
-    tagBackgroundColor:"#39C5BB"
+    tagBackgroundColor:"#39C5BB",
+    message: undefined,
   },
 
   tapName: function(event) {
     console.log(event)
   },
   
-  
   getActivityList: function() {
     let self = this,
         third_session = wx.getStorageSync('third_session');
     console.log('get third_session: ', third_session)
-    wx.getLocation({
-        success: res => {
-            self.setData({
-              longitude: res.longitude,
-              latitude: res.latitude
-            })
-        }
-    })
     let send = {
             third_session: third_session,
             limit: JSON.stringify(self.data.limit_per_request),
             lastActivityTime: JSON.stringify(self.data.lastActivityTime),
-            longitude: JSON.stringify(self.data.longitude),
-            latitude: JSON.stringify(self.data.latitude),
         }
-    console.log('explore sends ', send)
+    console.log('sends ', send)
     wx.request({
-        url: app.globalData.rootUrl + 'user/getActivityList',
+        url: app.globalData.rootUrl + 'user/UserActivityHistory',
         data: {
             third_session: third_session,
+            character: JSON.stringify("initiator"),
+            status: JSON.stringify("全部"),
             limit: JSON.stringify(self.data.limit_per_request),
             lastActivityTime: JSON.stringify(self.data.lastActivityTime),
-            longitude: JSON.stringify(self.data.longitude),
-            latitude: JSON.stringify(self.data.latitude),
         },
         method: 'POST',
         header: {
@@ -64,8 +50,7 @@ Page({
         },
         
         success:function(res){
-            console.log('request getActList returns: ', res.data)
-            console.log('request getActList returns: ', res.data.alist)
+            
             let oldList = self.data.fakeList,
                 newList = self.data.fakeList;
                 
@@ -82,13 +67,28 @@ Page({
             let list = oldList.concat(newList),
                 randomColorArr = self.data.randomColorArr.concat(util.generateRandomBgColor(newList.length));
             
+            /*
+            // For debug
+            if (list.length == 0)
+              list = self.data.newlist
+            */
+            // For deployment
+            // If current activity list is empty, show a message.
+            if (list.length == 0) {
+                self.setData({
+                    message: 'You have not initated any activity yet!',
+                    show_message:true
+                })
+            }
+            
+            util.generateBgColor(list)
+            
             self.setData({
                 activitylist: list,
                 lastActivityTime: res.data.lastActivityTime,
                 randomColorArr: randomColorArr
             })
             
-            console.log(self.data.activitylist)
         },
         fail: function(res) {
             console.log('登陆失败！' + res.errMsg)
@@ -107,60 +107,71 @@ Page({
         url: '../activityInfo/activityInfo?query=' + query,
       })
   },
-  
-  clearDataOnRefresh: function () {
-    var self = this;
-    self.setData({
-      activitylist: [],
-      lastActivityTime: '',
-    })
-  },
   /**
    * 生命周期函数--监听页面加载
    */
   onLoad: function (options) {
-    console.log('onLoad:')
-    var self = this;
-    self.clearDataOnRefresh()
+      console.log('onLoad:')
+      var self = this;
+      self.setData({
+          activitylist: [],
+          lastActivityTime: '',
+      })
     // 获取最新发布的活动列表
-    wx.getLocation({
-        success: res => {
-            self.setData({
-                longitude: res.longitude,
-                latitude: res.latitude
-            })
-            self.getActivityList()
-        }
-    })
+    self.getActivityList()
     
   },
 
+  /**
+   * 生命周期函数--监听页面初次渲染完成
+   */
+  onReady: function () {
+
+  },
 
   /**
    * 生命周期函数--监听页面显示
    */
   onShow: function () {
-    console.log('onShow:')
-    var self = this;
-    self.clearDataOnRefresh()
-    // 获取最新发布的活动列表
-    wx.getLocation({
-        success: res => {
-            self.setData({
-                longitude: res.longitude,
-                latitude: res.latitude
-            })
-            self.getActivityList()
-        }
-    })
+      console.log('onShow:')
+      /*
+      var self = this;
+      self.setData({
+          activitylist: [],
+          lastActivityTime: '',
+      })
+      self.getActivityList()
+      */
   },
 
+  /**
+   * 生命周期函数--监听页面隐藏
+   */
+  onHide: function () {
+
+  },
+
+  /**
+   * 生命周期函数--监听页面卸载
+   */
+  onUnload: function () {
+
+  },
 
   /**
    * 页面相关事件处理函数--监听用户下拉动作
    */
   onPullDownRefresh: function () {
       console.log('onPullDownRefresh')
+  },
+  upperHandler: function() {
+      console.log('upperHandler')
+    // TODO: 获取最新发布的活动列表
+    
+      self.setData({
+          activitylist: self.data.newlist
+      })
+      
   },
 
   /**
@@ -178,7 +189,7 @@ Page({
    */
   bottomHandler: function () {
       console.log('bottomHandler')
-      let self=this;
+      let self = this;
       self.getActivityList()
       console.log(self.data.activitylist)
   },
@@ -190,19 +201,42 @@ Page({
 
   },
   onRefreshClick: function() {
-    console.log('onRefreshClick:')
-    var self = this;
-    self.clearDataOnRefresh()
+      console.log('onRefreshClick:')
+      var self = this;
+      self.setData({
+          activitylist: [],
+          lastActivityTime: '',
+      })
     // 获取最新发布的活动列表
-    wx.getLocation({
-        success: res => {
-            self.setData({
-                longitude: res.longitude,
-                latitude: res.latitude
-            })
-            self.getActivityList()
+    self.getActivityList()
+      
+  },
+  
+  clickDeleteButton: function(e){
+    let self = this;
+    let index = e.currentTarget.dataset.actindex,
+        activity_id = self.data.activitylist[index].id;
+    console.log('clickDeleteButton: ', activity_id)
+    console.log('e: ', e)
+    wx.request({
+        url: app.globalData.rootUrl + "user/exitfromActivity",
+        data: {
+          third_session: wx.getStorageSync('third_session'),
+          activity_id: activity_id
+        },
+        method: "POST",
+        header: {
+          'content-type': 'application/x-www-form-urlencoded',
+          'chartset': 'utf-8'
+        },
+        success: function (res) {
+          console.log(res);
+          wx.showToast({
+            title: res.data,
+            icon: 'none',
+            duration: 4000
+          })
         }
     })
-      
   }
 })
