+++ pages/subPackage/participatedActivity/participatedActivity.js
@@ -1,6 +1,6 @@
 // pages/participatedActivity/participatedActivity.js
 const app = getApp();
-var util = require('../../utils/util.js');
+var util = require('../../../utils/util.js');
 Page({
 
   /**
