+++ pages/subPackage/participatedActivity/participatedActivity.wxml
