+++ pages/subPackage/participatedActivity/participatedActivity.wxss
